// src/models/initModels/nameModels.ts
export enum NameModels {
  antecedent = "antecedent",
  compteBloque = "compteBloque",
  consultation = "consultation",
  dossierPatient = "dossierPatient",
  historiqueMessage = "historiqueMessage",
  infoClinique = "infoClinique",
  message = "message",
  ordonnance = "ordonnance",
  patient = "patient",
  payement = "payement",
  personnel = "personnel",
  poste = "poste",
  rendezVous = "rendezVous",
  salle = "salle",
  tache = "tache",
  typeDeSalle = "typeDeSalle",
  typeRendezVous = "typeRendezVous",
}